+++ src/CatX/App.xaml.cs
@@ -2,6 +2,6 @@ using System.Windows;
 
 namespace CatX;
 
-public partial class App : Application
+public partial class App : System.Windows.Application
 {
 }